api/requirements.txt
Flask==1.0.3
pytest==4.6.3
pymongo==3.8.0
py2neo==4.3.0
python-logstash-async==1.5.1
python-logstash==0.4.6
flask-swagger-ui==3.20.9
